refactor: Calculate matches outside of build_directory_tree_data
Previously, the regex was being tested within the function
`build_directory_tree_data`, but the problem with that method is that
extra work needs to be done to determine whether a subtree contains the
match. Doing all of this work within the much simpler
`mark_matched_nodes` function is a lot cleaner.