refactor: Rename some functions
`consume_7e` is now `handle_nav_key`, which is a less obscure name.

`draw` is now `render` because it is the entrypoint of the `render`
module.

`render_directory_tree` is now `flatten_tree` because
`render_directory_tree` suggests side-effects, but actually we are just
turning a tree data structure into a vector data structure.